Refactoring the initial structure of the files
